pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fnerd-falconpy"
version = "1.3.0"
description = "fnerd-falconpy - Forensic collection toolkit for CrowdStrike Falcon RTR"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "fnerd", email = "fnerd@example.com"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Security",
]
keywords = ["crowdstrike", "falcon", "rtr", "forensics", "incident-response", "fnerd", "falconpy"]

dependencies = [
    "crowdstrike-falconpy>=1.3.0",
    "boto3>=1.26.0",
    "python-dotenv>=0.19.0",
    "requests>=2.28.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.990",
    "pylint>=2.15.0",
]

[project.scripts]
fnerd-falconpy = "fnerd_falconpy.cli.main:main"

[project.urls]
Homepage = "https://github.com/fnerd/fnerd-falconpy"
Documentation = "https://github.com/fnerd/fnerd-falconpy/blob/main/README.md"
Repository = "https://github.com/fnerd/fnerd-falconpy"
Issues = "https://github.com/fnerd/fnerd-falconpy/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["fnerd_falconpy*"]

[tool.setuptools.package-data]
"fnerd_falconpy.resources" = ["**/*"]